First step towards external tests
Tests are now done in C, or other languages. Or will be.
This is the first step towards this goal.
Tests are standalone binaries, whic communicate with Insanity
via D-Bus. A sample C test (which does nothing, and which is
still on the quick and dirty side) is included for reference.
Most of the code in this test will be made into a library so
that future tests can just supply their implementation of
three calls (setup, test, stop). There will be a layer above
this that will take care of generic GStreamer pipeline as well.