Split away the lib code
Now that we have something that works as a proof of concept,
move out the generic code into a separate reusable lib, and
make the blank test use it.

Extra layers (eg, a GStreamer layer) should fit here by plugging
in the hooks, and letting the user plug into their own hooks.

The Makefile's a hand made hardcoded one for now, not sure what
the build system may end up being if we're to run on Windows too.