Remove some DBus signals and rework the test startup process
There are no ready/stop signals anymore, only a done signal.

Tests are supposed to be started if the setup DBus method returned
success (test is ready). And tests are supposed to be stopped when
the done signal happened and another iteration is only to be started
when the stop DBus method gave a reply.