dbstorage: Some syntax fixes to make db merging work again
__mergeTestRun: Fixed getting the parent id for updating

the newly inserted tests. It was looking for 3 arguments
when only 2 are needed for the update statement, the new
test id and the parent id.

__mergeTest: monitorclassmap argument is not used. Removed
 from the function declaration

__rawInsertTestClassInfo: This function changed the
arguments order. Fix the calls to make the call by name
instead of by order, to prevent from future failures

__rawInsertTestClassInfo: If the parent id is blank, make
sure the inserted id is an empty string, not an empty
dictionary, which was throwing an error